+++ src/builtins/filters/number.rs
@@ -1,6 +1,7 @@
 /// Filters operating on numbers
 use std::collections::HashMap;
 
+#[cfg(feature = "builtins")]
 use humansize::{file_size_opts, FileSize};
 use serde_json::value::{to_value, Value};
 
@@ -59,6 +60,7 @@ pub fn round(value: &Value, args: &HashMap<String, Value>) -> Result<Value> {
 }
 
 /// Returns a human-readable file size (i.e. '110 MB') from an integer
+#[cfg(feature = "builtins")]
 pub fn filesizeformat(value: &Value, _: &HashMap<String, Value>) -> Result<Value> {
     let num = try_get_value!("filesizeformat", "value", i64, value);
     num.file_size(file_size_opts::CONVENTIONAL)
@@ -171,6 +173,7 @@ mod tests {
         assert_eq!(result.unwrap(), to_value(2.9).unwrap());
     }
 
+    #[cfg(feature = "builtins")]
     #[test]
     fn test_filesizeformat() {
         let args = HashMap::new();